friend and game share ui
